+++ src/hooks/useGetBlockCountdown.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from 'react'
 import { BSC_BLOCK_TIME } from 'config'
-import { simpleRpcProvider } from 'utils/web3'
+import { simpleRpcProvider } from 'utils/providers'
 
 /**
  * Returns a countdown in seconds of a given block
