+++ src/state/hooks.ts
@@ -7,7 +7,7 @@ import { orderBy } from 'lodash'
 import { Team } from 'config/constants/types'
 import Nfts from 'config/constants/nfts'
 import { farmsConfig } from 'config/constants'
-import { simpleRpcProvider } from 'utils/web3'
+import { simpleRpcProvider } from 'utils/providers'
 import { getBalanceAmount } from 'utils/formatBalance'
 import { BIG_ZERO } from 'utils/bigNumber'
 import useRefresh from 'hooks/useRefresh'
